fix(viewer-lib): Renamed OrientedSectionTool back to SectionTool since we not going to be doing multiple section tool types
